feat!: freertos support implemented

+++ include/zh_avr_vector.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "FreeRTOS.h"
+#include "task.h"
 #include "stdlib.h"
 #include "stdint.h"
 #include "string.h"
@@ -30,11 +32,11 @@ extern "C"
 
     typedef struct // Main structure of vector data.
     {
-        void **items;      // Array of pointers of vector items.
-        uint16_t capacity; // Maximum capacity of the vector. @note Used to control the size of allocated memory for array of pointers of vector items. Usually equal to the current number of items in the vector. Automatically changes when items are added or deleted.
-        uint16_t size;     // Number of items in the vector. @note Can be read with zh_avr_vector_get_size().
-        uint16_t unit;     // Vector item size. @note Possible values from 1 to 65535.
-        bool status;       // Vector initialization status flag. @note Used to prevent execution of vector functions without prior vector initialization.
+        void **items;     // Array of pointers of vector items.
+        uint8_t capacity; // Maximum capacity of the vector. @note Used to control the size of allocated memory for array of pointers of vector items. Usually equal to the current number of items in the vector. Automatically changes when items are added or deleted.
+        uint8_t size;     // Number of items in the vector. @note Can be read with zh_avr_vector_get_size().
+        uint8_t unit;     // Vector item size. @note Possible values from 1 to 255.
+        bool status;      // Vector initialization status flag. @note Used to prevent execution of vector functions without prior vector initialization.
     } zh_avr_vector_t;
 
     /**
@@ -43,22 +45,16 @@ extern "C"
      * @param[in] vector Pointer to main structure of vector data.
      * @param[in] unit Size of vector item.
      *
-     * @return
-     *              - AVR_OK if initialization was success
-     *              - AVR_ERR_INVALID_ARG if parameter error
-     *              - AVR_ERR_INVALID_STATE if vector already initialized
+     * @return AVR_OK if success or an error code otherwise.
      */
-    avr_err_t zh_avr_vector_init(zh_avr_vector_t *vector, uint16_t unit);
+    avr_err_t zh_avr_vector_init(zh_avr_vector_t *vector, uint8_t unit);
 
     /**
      * @brief Deinitialize vector. Free all allocated memory.
      *
      * @param[in] vector Pointer to main structure of vector data.
      *
-     * @return
-     *              - AVR_OK if deinitialization was success
-     *              - AVR_ERR_INVALID_ARG if parameter error
-     *              - AVR_ERR_INVALID_STATE if vector not initialized
+     * @return AVR_OK if success or an error code otherwise.
      */
     avr_err_t zh_avr_vector_free(zh_avr_vector_t *vector);
 
@@ -67,9 +63,7 @@ extern "C"
      *
      * @param[in] vector Pointer to main structure of vector data.
      *
-     * @return
-     *              - Vector size
-     *              - AVR_FAIL if parameter error or vector not initialized
+     * @return AVR_OK if success or an error code otherwise.
      */
     avr_err_t zh_avr_vector_get_size(zh_avr_vector_t *vector);
 
@@ -79,11 +73,7 @@ extern "C"
      * @param[in] vector Pointer to main structure of vector data.
      * @param[in] item Pointer to item for add.
      *
-     * @return
-     *              - AVR_OK if add was success
-     *              - AVR_ERR_INVALID_ARG if parameter error
-     *              - AVR_ERR_NO_MEM if memory allocation fail or no free memory in the heap
-     *              - AVR_ERR_INVALID_STATE if vector not initialized
+     * @return AVR_OK if success or an error code otherwise.
      */
     avr_err_t zh_avr_vector_push_back(zh_avr_vector_t *vector, void *item);
 
@@ -94,13 +84,9 @@ extern "C"
      * @param[in] index Index of item for change.
      * @param[in] item Pointer to new data of item.
      *
-     * @return
-     *              - AVR_OK if change was success
-     *              - AVR_ERR_INVALID_ARG if parameter error
-     *              - AVR_ERR_INVALID_STATE if vector not initialized
-     *              - AVR_FAIL if index does not exist
+     * @return AVR_OK if success or an error code otherwise.
      */
-    avr_err_t zh_avr_vector_change_item(zh_avr_vector_t *vector, uint16_t index, void *item);
+    avr_err_t zh_avr_vector_change_item(zh_avr_vector_t *vector, uint8_t index, void *item);
 
     /**
      * @brief Get item by index.
@@ -108,11 +94,9 @@ extern "C"
      * @param[in] vector Pointer to main structure of vector data.
      * @param[in] index Index of item for get.
      *
-     * @return
-     *              - Pointer to item
-     *              - NULL if parameter error or vector not initialized or if index does not exist
+     * @return AVR_OK if success or an error code otherwise.
      */
-    void *zh_avr_vector_get_item(zh_avr_vector_t *vector, uint16_t index);
+    void *zh_avr_vector_get_item(zh_avr_vector_t *vector, uint8_t index);
 
     /**
      * @brief Delete item by index and shifts all elements in vector.
@@ -120,13 +104,9 @@ extern "C"
      * @param[in] vector Pointer to main structure of vector data.
      * @param[in] index Index of item for delete.
      *
-     * @return
-     *              - AVR_OK if delete was success
-     *              - AVR_ERR_INVALID_ARG if parameter error
-     *              - AVR_ERR_INVALID_STATE if vector not initialized
-     *              - AVR_FAIL if index does not exist
+     * @return AVR_OK if success or an error code otherwise.
      */
-    avr_err_t zh_avr_vector_delete_item(zh_avr_vector_t *vector, uint16_t index);
+    avr_err_t zh_avr_vector_delete_item(zh_avr_vector_t *vector, uint8_t index);
 
 #ifdef __cplusplus
 }
